refactor(ui): replace all QML plugins
Having multiple QML plugins in the same static-linked application causes linker errors.
We must allow static builds because WASM may only link statically.
Therefore, plugins cannot be used.

Now, each library must provide a "registration" file with a registerTypes method.
This method is responsible for QML<=>C++ mapping.
