--- ui/memory/plugin.cpp
@@ -1,3 +1,0 @@
-#include "plugin.hpp"
-
-void Memory::registerTypes(const char *uri) {}
